requirements.txt
python-dotenv==0.19.2
mysql-connector-python
bs4
markdownify